% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_FSBT.R
\name{emp_FSBT}
\alias{emp_FSBT}
\title{Full Significance Bayesian Testing}
\usage{
emp_FSBT(H0, sample)
}
\arguments{
\item{H0}{The value under the null hypothesis.}

\item{sample}{A monte Carlo sample from the marginal posterior distribution.}
}
\value{
double
}
\description{
Performs Full Significance Bayesian Testing (FSBT) for univariate sharp
null hypothesis based on a posterior sample. The marginal posterior density is obtained by kernel density estimation from the posterior sample provided through the \code{sample} argument.
}
\references{
Pereira, C. A. d. B., Stern, J. M. and Wechsler, S. (2008) Can a significance test be genuinely Bayesian? \emph{Baysian Analysis} \bold{3}, 79-100.
}
