% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_hpd.R
\name{emp_hpd}
\alias{emp_hpd}
\title{Empirical Highest Posterior Density Interval}
\usage{
emp_hpd(x, prob = 0.95)
}
\arguments{
\item{x}{a univariata or marginal posterior sample.}

\item{prob}{the required posterior probability content.}
}
\description{
Computes empirical Highest Posterior Density (HPD) interval from a posterior sample. Works only for scalar marginal posteriors.
}
