% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_hpd.R
\name{hpd}
\alias{hpd}
\title{Empirical Highest Posterior Density Interval}
\usage{
hpd(x, prob = 0.95)
}
\arguments{
\item{x}{a univariate or a posterior sample for a scalar parameter.}

\item{prob}{posterior probability content.}
}
\value{
vector of two doubles.
}
\description{
Computes the Highest Posterior Density (HPD) interval from a posterior sample. Works only for scalar marginal posteriors.
}
