% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_h.R
\name{tune.h}
\alias{tune.h}
\title{Tune ABC distance bandwidth}
\usage{
tune.h(formula, data, ..., n.samp = 1e+5, n.sim.HJ = 500, acc.rate, grid.h, prior,
       cHub = 1.345, cHub2 = 2.07,
       init, n.cores = 1, use.h)
}
\arguments{
\item{formula}{a two-sided linear formula object describing the
fixed-effects part of the model, with the response on the left of
a \code{~} operator and the terms, separated by \code{+}
operators on the right.  The \code{"|"} character separates an
expression for a model matrix and a grouping factor.}

\item{data}{an optional data frame containing the variables named
in \code{formula}. By default the variables are taken from the
environment of \code{lmer} called internally.}

\item{...}{other aruments to be passed to lmer(). Currently none is used.}

\item{n.samp}{the number of pilot posterior samples to be drawn with ABC for each value of \code{grid.h}.}

\item{n.sim.HJ}{the number of simulations to be used for computing the sensitivity and variabiliy matrices.}

\item{acc.rate}{the desired acceptance rate of the ABC-MCMC algorithm.}

\item{grid.h}{a grid of \eqn{h} values within which the "optimal" value is to be found.}

\item{prior}{A named list of user-defined prior hyper-parameters. See "Details" below.}

\item{cHub}{The tunnin constant of the Huber function for the location parameter.}

\item{cHub2}{The tunnin constant of the Huber proposal 2 function for the scale parameter.}

\item{init}{optional object to use for starting values. Currently ignored as initial values are taken from \code{lmer}.}

\item{n.cores}{the number of cores for parallel computation.}

\item{use.h}{a bandwidth to be used for the ABC distance. If provided, no tuning for \eqn{h} is performed and \code{acc.rate} argument is ingored.}
}
\value{
a list.
}
\description{
Tunes the bandwidth \eqn{h} of the ABC distance to get the desired level of acceptance rate specified via \code{acc.rate}. Besides tuning \eqn{h}, the function also gets the relevant quantities need for running ABC-MCMC by a preliminary run of the \code{lmer} command.
}
\details{
Given a specifiction of the \code{formula} and \code{data} the function calls internally \code{rlmer} to get the REML estimates and extracts from the resulting object the necesary quantites. Then proceeds by finding the solution of the REML II robust estimating equations and computes the sensitiviy and the variability matrices. Finally for each value of code{grid.h} draws \code{n.samp} posterior samples with the ABC-MCMC algorithm and saves the acceptance rate. Finally a functions is built by a smoothing spline the acceptance rates vs \code{grid.h}. The "optimal" value of \eqn{h} in \code{grid.h} is found by inverting the spline function at \code{acc.rate}. Currently, the prior for the fixed effects is the product of scalar normals with mean zero and user-specified variance. All fixed parameters are assumed to have equal prior variance. For the variance components the prior is halfCauchy with user-specified scale. Both variance parameters are assumed to have equal prior scale.
}
